Use a more general definition of the mutator-functions
Make the mutator functions signature like &[GenomeType] -> GenomeType.
This makes it possible to make mutators and crossover functions with the same
signature, and also makes it posible (with a little bit of inconveinence) for
the user, to select what genome to mutate, or to select parents for a child
genome. This signature change might make the library function a little more complicated to use, as the selection was previously done by the library.